feat: add support of simple types to AST
For now:

- Named types: eg. `number`
- Generic types: eg. `Array<T>`
- Tuple types: eg. `[K, V]`
- Union types: eg. `string | null`

As evidenced in this commit, you can created nested types (ie. with
internal structure) like `Array<[Production | Token | null, number]>`.

Like everything in this project, more could be added, but will only do
that as needed.

Note that I needed to get creative with the grammar to avoid ambiguity:

    Type → NamedType
    Type → GenericType
    Type → UnionType
    Type → TupleType

Can't use a definition like:

    UnionType → TypeVariantList
    TypeVariantList → Type
    TypeVariantList → TypeVariantList BITWISE_OR Type
    TypeVariantList → ε

Because this is ambiguous (a single `Type` could be a `UnionType` or any
other `Type`). I make the simplifying assumptions that:

- We don't support leading BITWISE_OR (yet... we could quite easily I
  think).
- We don't support empty tuples.
- We assume left associativity and verify that precedence works (see
  tests), but we don't worry about "weird" types like nested unions made
  with parentheses (these should end up flattened anyway).