refactor(parser): remove TODO comment
I checked this out, replacing this:

    const popped: Array<Production | Token | null> = [];
    for (let i = 0; i < pop; i++) {
      const [node] = stack.pop()!;
      popped[pop - i - 1] = node;
    }

with this:

    const popped: Array<Production | Token | null> = stack
      .slice(-pop)
      .map(([node]) => node);
    stack.length = stack.length - pop;

Motivation here is twofold:

1. See if it is faster (ie. instead of N `pop()` calls, do one `slice()`
   and one `map()`.
2. See if I can make something that will be easier to parse.

The verdict is in on point number 1; it's not faster...

Before:

|         |             duration |
|---------|----------------------|
| warm-up |          50.014334ms |
|     run | 32.398875000000004ms |

After:

|         |            duration |
|---------|---------------------|
| warm-up |         97.253166ms |
|     run | 77.79612500000002ms |

And as for number 2, I think there's no big difference. So, removing the
TODO comment as I won't be proceeding with this idea, at least not in
this form. An alternative might be to use `splice()` instead of
`slice()` plus `length`, but that still wouldn't save us the `map()`
call, which is probably what's slowing us down.