fix(typescript): parse internal DocComment structure
The semantic action is horrible and it might be better to do this in the
lexer, but anyway.

Note the use of the `flatMap` trick as described here:

- https://stackoverflow.com/a/70738075/2103996

This is because I can't use types in semantic actions (because they are
`RawStatement` and we can't `walk()` into those in a meaningful way) and
TypeScript wrongly infers the type to be `string | null` if we write:

    lines.map((line) => {
      // Code that may return a string or `null`.
    }).filter((line) => line !== null)

It _can_ infer this kind of thing though, I think, or at least I seem to
recall having it work in other places:

    lines.map((line) => {
      // Code that may return a string or `null`.
    }).filter(Boolean)

but that is no good for our purposes here (because as the tests
illustrate, we want to be able to include empty strings in the result,
and those are falsey and would be filtered).