chore: fix new TS errors caused by stricter compiler
New version of TS is complaining about `string` types not matching
concrete literals. For example, formatted for readability:

    packages/typescript/src/__tests__/parseStatement.test.ts:81:59 -
    error TS2322: Type '{
      kind: string;
      expression: {
        kind: string;
        arguments: never[];
        callee: {
          kind: string;
          name: string;
        };
      };
    }' is not assignable to type 'Statement'.

There were about 7 other variants of that message, but that is one of
the most readable ones and illustrates the point well enough. An `as
const` isn't enough to make the problem go away; I had to explicitly tag
the values at initialization with the exact type.

No idea why the `never` comes in in the example above. A different
example without that weird anomaly would be:

    packages/lexer/src/RegExp/__tests__/RegExpParser.test.ts:63:58 -
    error TS2345: Argument of type '{
      kind: string;
      children: {
        kind: string;
        from: string;
        to: string;
      }[];
      negated: boolean;
    }' is not assignable to parameter of type 'Atom | CharacterClass |
    Range | Alternate | Anything | Repeat | Sequence'.