docs: add Redis cheatsheet
Because every time I look at the `updateIndices.js` script in the
current implementation I have to squint at it to remember exactly how
Redis is being used. As noted here:

- https://github.com/wincent/masochist/issues/218

I'll probably be using Redis to cache more things in the future, so I
want to have a good handle on what and how it's being used before going
down that road.