feat: start fleshing out new indexer
I'm committing this in unfinished form because I realized I'm going down
the wrong track and I want the Git history to tell the tale of my sins.

Just because `Bun.spawn` returns `ReadableStream`s for the I/O doesn't
mean you have to use them. Look at all the book-keeping I am about to
create just to read this stuff in in chunks. It's going to be faster to
read it all in at once with `text()` instead, and avoid the intermediate
strings that would come via `split()` by using `indexOf()`.
`ReadableStream`s are probably the right choice for slow stuff coming in
off the network, but that's not the case here; it's gonna come in as
fast as Git can write it out, and that will be plenty fast.