feat: put tags into a useful datastructure
Note `export {}` may be temporary, but needed for now because `bun tsc`
prints:

```
TS1375: 'await' expressions are only allowed at the top level of a file
when that file is a module, but this file has no imports or exports.
Consider adding an empty 'export {}' to make this file a module.
```