perf: read contents from Git
On the server we have a bare repo, so can't read the blobs in with
`Bun.file`. But we can get `git-ls-files` to print blob IDs for us
instead of just path names, and we can pipe those blob IDs into
`git-cat-file` in `--batch` mode and have it read out the blob contents
for us in a single fork, so the perf should be great.

As noted in the comments, we're very much assuming well-formed input
here, but I think that's ok as we control the inputs.