fix: don't die for internal redirect
For an article like `/wiki/Issue_tracker`, which redirects to
`/issues/`, we were:

- In `ArticleRoute`, calling `softRedirect()`.
- Which checks to see if it is a valid route (it is, because it
  satisfies the catch-all), and calls `makeInternalRedirect()`.
- `makeInternalRedirect` was creating an `InternalRedirectError` with a
  `target` of `/issues/`.
- in `src/server/main.js`, we were calling `resolve(error.target)`.
- `resolve()` was in turn calling `router.resolve({pathname:
  location.pathname})` (note that `target` is just a string, so
  `location.pathname` was effectively `undefined`).
- inside universal-router, it ends up calling `substr()` on this
  pathname and blowing up.

Pretty sure this used to work, so not sure when things broke.

Note that this commit fixes the 500 errors, but most internal redirects
are broken in dev because we're not running nginx. I think that in prod
they will work, because nginx will serve up static files if it can
before hitting the app.